requirements.txt
tensorflow==2.15.0
tensorflow-addons==0.21.0
tensorflow-model-optimization==0.8.0
numpy==1.24.4
pandas==2.1.2
scikit-learn==1.3.2
matplotlib==3.8.0
opencv-python==4.8.1.78
streamlit==1.29.0
streamlit-webrtc==0.47.3
python-dotenv==1.0.0
tqdm==4.66.1
Pillow==10.1.0
seaborn==0.13.0
pytest==7.4.3
pytest-mock==3.11.1
absl-py==1.4.0
protobuf==3.20.3
mediapipe==0.10.8
